requirements.txt
streamlit==0.79.0
pandas==1.2.1
plotly==4.14.3
Pillow==8.1.0
requests==2.25.1